Move codgen scripts and files into generator folder
Including moving vk_dispatch_table_helper.h into its own folder.

Allows for removing all of the verify_exclude filenames since the generator
folders only contain generated files.
