+++ scripts/generators/dispatch_table_helper_generator.py
