+++ scripts/generators/helper_file_generator.py
