+++ scripts/generators/loader_extension_generator.py
